refactor: move authentication domain to users domain
