--- hals/src/users/user.ts
@@ -1,5 +1,0 @@
-export type User = {
-   _id : string,
-   username : string,
-   hash : string,
-};
